ov091: promote daDsnBase_c into a single C++ TU (9 of 11)
Gathers nine of the eleven one-function shards in ov091's 0x02132d04..
0x02133254 run into src/actors/daDsnBase_c.cpp. The class already had a
real header; this is gathering, not authoring.

  tubuild verify   9/9 MATCH, objisolate clean, reloc-destinations clean,
                   emitted .text ROM-ascending
  tubuild linkcheck (pre-promotion) SCRATCH-LINK-VERIFIED -- [4b/8] 9
                   LICENSED / 0 HOMELESS / 0 COLLIDES, the whole range
                   0x02132dc0..0x02133254 IDENTICAL from one object,
                   module fidelity 106/106 exact (100.000%), ROM sha256
                   d1506e90... identical to the stock control. `dsd check
                   symbols --fail` still fails on the same 9 pre-existing
                   errors the baseline control has: 0 new, 0 resolved.

WHY 9 AND NOT 11. The cartridge orders D0 (0x02132d04) BELOW D1
(0x02132d6c). mwccarm 2004/b56's destructor-variant lever set is closed: a
destructor defined inline in the class body emits D1 then D0, always; an
out-of-line one emits D2, D0, D1 and adds a homeless D2 the ROM does not
carry. Scaffolds and declaration position change nothing. Both variants
BYTE-MATCH from the inline form -- verify reported 11/11 before the range
was narrowed -- so the source form is right and only the ORDER is
unreproducible, and linkcheck [4b/8] fatally refuses a licensed .text that
is not ROM-ascending.

So the two destructor shards stay enrolled and this TU's own copies are
licensed deadstrip-duplicate. Nothing outside the vtable depends on those
addresses: the only references to either in all of config/ are relocs.txt
0x0213523c -> 0x02132d6c and 0x02135240 -> 0x02132d04, i.e. vtable slots 16
and 17. No instruction in the image calls either address.

THE FORCING SCAFFOLDS ARE GONE. Both destructor shards carried one
(`p->~daDsnBase_c();`, `delete p;`). This TU defines the key function
(CleanupResources -- the first declared virtual that is neither inline nor
pure), and that alone emits the vtable and both variants byte-identically.

COMPILER-ONLY OUTPUT: 15 rows, externalized_output EMPTY. Ten inherited and
own _ZTI/_ZTS records plus _ZTV11daDsnBase_c are deadstrip-data; the D0/D1
pair and _ZN7Vector3D1Ev are deadstrip-duplicate; only the file-local
anonymous struct's D1 is genuinely homeless and takes plain deadstrip. The
RTTI record is homed in ov025 and the vtable in ov091 -- a cross-module home
is still a home, so it is deadstrip-data with canonical_module/address
pointing where it actually lives, not externalized.

RECONCILIATION. decl_common.h:2858 already declared func_ov091_02133098
taking char* where the shard defined void*; the merge made both visible and
mwccarm rejected it as `illegal function overloading`, so the real header
won. The shards' shadow `struct dActor_c` and `struct Sub` gave way to the
real dActor_c::OnAimedAtWithEgg (slot 29) and Model::Render (slot 5). Both
tubuild CONFLICT notes are resolved and their markers deleted from the
source. Include order is load-bearing and commented as such: daDsnBase_c.h
reaches dBgActor_c.h, which includes common.h before Model.h, fixing
Matrix4x3 to the flat spelling every shard compiled against.
