+++ config/converted-backslide-exceptions.jsonl
@@ -295,3 +295,4 @@
 {"path": "src/_ZN10dBgActor_cC2Ev.cpp", "reason": "TU promotion of ov002/dBgActor_c: byte-match required keeping raw offset arithmetic. Reconciling *(short*)((char*)this + 0x8c) to the named angle fields let mwcc CSE the field address and dropped 4 instructions from UpdateKillByMegaChar (0x158 -> 0x148); the raw spelling was restored on exactly those accesses to hold 11/11 MATCH. Remaining member failures (mangled _Z call spellings, the still-unnamed func_ov002_020ee5d0, residual unk_ fields) are reconcilable readability work that does not affect bytes, and are left for a follow-up pass rather than risking the match in a promotion PR."}
 {"path": "src/_ZN10dBgActor_cD0Ev.cpp", "reason": "TU promotion of ov002/dBgActor_c: byte-match required keeping raw offset arithmetic. Reconciling *(short*)((char*)this + 0x8c) to the named angle fields let mwcc CSE the field address and dropped 4 instructions from UpdateKillByMegaChar (0x158 -> 0x148); the raw spelling was restored on exactly those accesses to hold 11/11 MATCH. Remaining member failures (mangled _Z call spellings, the still-unnamed func_ov002_020ee5d0, residual unk_ fields) are reconcilable readability work that does not affect bytes, and are left for a follow-up pass rather than risking the match in a promotion PR."}
 {"path": "src/_ZN10dBgActor_cD1Ev.cpp", "reason": "TU promotion of ov002/dBgActor_c: byte-match required keeping raw offset arithmetic. Reconciling *(short*)((char*)this + 0x8c) to the named angle fields let mwcc CSE the field address and dropped 4 instructions from UpdateKillByMegaChar (0x158 -> 0x148); the raw spelling was restored on exactly those accesses to hold 11/11 MATCH. Remaining member failures (mangled _Z call spellings, the still-unnamed func_ov002_020ee5d0, residual unk_ fields) are reconcilable readability work that does not affect bytes, and are left for a follow-up pass rather than risking the match in a promotion PR."}
+{"path": "src/_ZN11daDsnBase_c16CleanupResourcesEv.cpp", "reason": "ov091/daDsnBase_c: partial TU promotion, 9 of 11 shards absorbed into src/actors/daDsnBase_c.cpp (licensed .text 0x02132dc0-0x02133254). The D0/D1 pair stays enrolled because the cartridge orders D0 below D1, which no admissible source form reproduces."}
