+++ config/converted-baseline.json
@@ -519,7 +519,6 @@
     "src/_ZN11dScMiniGm_c16OnPendingDestroyEv.cpp",
     "src/_ZN11dScMiniGm_cD0Ev.cpp",
     "src/_ZN11dScMiniGm_cD1Ev.cpp",
-    "src/_ZN11daDsnBase_c16CleanupResourcesEv.cpp",
     "src/_ZN11daDsnBase_cD0Ev.cpp",
     "src/_ZN11daDsnBase_cD1Ev.cpp",
     "src/_ZN11fLiNdBaPr_cD1Ev.cpp",
@@ -2185,6 +2184,7 @@
     "src/actors/dBgActor_c.cpp#_ZN10dBgActor_cC2Ev",
     "src/actors/dBgActor_c.cpp#_ZN10dBgActor_cD0Ev",
     "src/actors/dBgActor_c.cpp#_ZN10dBgActor_cD1Ev",
+    "src/actors/daDsnBase_c.cpp#_ZN11daDsnBase_c16CleanupResourcesEv",
     "src/engine/fader/_ZN10FaderColorD0Ev.cpp",
     "src/engine/fader/_ZN10FaderColorD1Ev.cpp",
     "src/engine/fader/_ZN15FaderBrightness10SetToStartEv.cpp",
