--- src/func_ov091_02132dc0.cpp
@@ -1,38 +1,0 @@
-//cpp
-// @symbol func_ov091_02132dc0
-/* recovered: shared common types */
-#include "common.h"
-#include "dBgW.h"
-
-struct dActor_c {
-    virtual void f00(); virtual void f01(); virtual void f02(); virtual void f03();
-    virtual void f04(); virtual void f05(); virtual void f06(); virtual void f07();
-    virtual void f08(); virtual void f09(); virtual void f10(); virtual void f11();
-    virtual void f12(); virtual void f13(); virtual void f14(); virtual void f15();
-    virtual void f16(); virtual void f17(); virtual void f18(); virtual void f19();
-    virtual void f20(); virtual void f21(); virtual void f22(); virtual void f23();
-    virtual void f24(); virtual void f25(); virtual void f26(); virtual void f27();
-    virtual void f28(); virtual int f29();
-};
-extern "C" {
-extern void* _ZN8dActor_c18ClosestWithActorIDEj(void*, unsigned int);
-extern int Vec3_Dist(void*, void*);
-
-int func_ov091_02132dc0(char* c) {
-    void* p = _ZN8dActor_c18ClosestWithActorIDEj(c, 9);
-    if (p != 0) {
-        volatile struct Vector3 v;
-        v.x = *(int*)(c + 0x5c);
-        v.y = *(int*)(c + 0x60);
-        v.z = *(int*)(c + 0x64);
-        v.y = v.y + ((dActor_c*)c)->f29();
-        if (Vec3_Dist((char*)c + 0x5c, (char*)p + 0x5c) < (*(int*)(c + 0xb8) << 3)) {
-            if (!((dBgW *)(c + 0x124))->IsEnabled()) {
-                ((dBgW *)(c + 0x124))->Enable((dActor_c *)(c));
-                return 1;
-            }
-        }
-    }
-    return 0;
-}
-}
